Fix 3177 array task.

+++ informatics.mccme.ru/Изучение языка программирования/Массивы/Одномерные массивы/Условия задач (для языка Питон)/Z. Большой сдвиг/3177.cpp
@@ -21,7 +21,7 @@ int main ()
 
 	int k;
 	cin >> k;
-	k %= arr.size ();
+	k %= (int) arr.size ();
 
 	rotate (arr.begin (), ((k > 0) ? arr.end () : arr.begin ()) - k, arr.end ());
